chore(docker): add docker and docker-compose configuration

+++ src/main/resources/db/migration/V1__create_tb_cars.sql
@@ -4,8 +4,8 @@ CREATE TABLE tb_cars(
     brand VARCHAR(100),
     body VARCHAR(100),
     manufactureYear INTEGER,
-    valuecar DOUBLE,
-    consume DOUBLE,
+    valuecar DOUBLE PRECISION,
+    consume DOUBLE PRECISION,
     description VARCHAR(255),
     imgurl VARCHAR(255)
 );
